requirements.txt
langgraph
tavily-api  # replace with actual package name if different
openai
python-dotenv
notebook